refactor(#331): centralize panel C types and remove infra logging
## Blocker 1 — Panel C 도메인 타입 중앙화 (MISTAKES.md Architecture #1)
- 레거시 Signal/SignalType (analysis 패널용) 을 AnalysisSignal/AnalysisSignalType 으로 rename 하여 이름 충돌 해소
- Panel C 타입 (Signal, TrendState, StockSignalResult, SectorSignalsResult 등) 을 domain/types.ts 로 이동
- domain/signals/types.ts 파일 삭제
- 전체 import 경로 @/domain/signals/types → @/domain/types 로 변경

## Blocker 2 — infrastructure console.warn 제거 (MISTAKES.md Infrastructure #3)
sectorSignalsApi.ts 의 console.warn 3곳 (캐시 읽기/쓰기 실패, 종목 fetch 실패) 모두 제거. graceful degradation 유지 (catch {} + 설명 주석).

## Suggestion — SectorSignalPanel quadrants useMemo 주석 보강
CONVENTIONS.md 의 reduce+spread O(N*M) 예외를 명시적으로 인용, 섹터당 ~8 종목 × 2-3 signals 범위에서 로컬 변이의 정당성 명확화.

검증: 100% 도메인/인프라 커버리지 유지, 1591 tests pass.
