+++ src/components/dashboard/SignalStockCard.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { cn } from '@/lib/cn';
-import type { StockSignalResult } from '@/domain/signals/types';
+import type { StockSignalResult } from '@/domain/types';
 import { SignalBadge } from './SignalBadge';
 
 interface SignalStockCardProps {
